refactor: Remove swagger api project
