+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.API/Program.cs
@@ -12,13 +12,13 @@ using xxENSONOxx.xxSTACKSxx.API.Authentication;
 using xxENSONOxx.xxSTACKSxx.API.Authorization;
 using xxENSONOxx.xxSTACKSxx.API.Models.Requests;
 using xxENSONOxx.xxSTACKSxx.Shared.API.Middleware;
-using xxENSONOxx.xxSTACKSxx.Shared.API.Swagger.Filters;
 using Microsoft.OpenApi.Models;
 using OpenTelemetry.Exporter;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using xxENSONOxx.xxSTACKSxx.API;
+using xxENSONOxx.xxSTACKSxx.API.Filters;
 using xxENSONOxx.xxSTACKSxx.Infrastructure;
 
 var builder = WebApplication.CreateBuilder(args);
