--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.API.Swagger/Filters/BasePathFilter.cs
@@ -1,24 +1,0 @@
-using System.Collections.Generic;
-using Microsoft.OpenApi.Models;
-using Swashbuckle.AspNetCore.SwaggerGen;
-
-namespace xxENSONOxx.xxSTACKSxx.Shared.API.Swagger.Filters
-{
-    public class BasePathFilter(string basePath) : IDocumentFilter
-    {
-        public string BasePath { get; } = basePath;
-
-        public void Apply(OpenApiDocument swaggerDoc, DocumentFilterContext context)
-        {
-            if (string.IsNullOrWhiteSpace(BasePath))
-            {
-                return;
-            }
-
-            swaggerDoc.Servers = new List<OpenApiServer>
-            {
-                new OpenApiServer {Url = BasePath}
-            };
-        }
-    }
-}
