--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.API.Swagger/xxENSONOxx.xxSTACKSxx.Shared.API.Swagger.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-    <Description>API Swagger package for common features used throughout the projects.</Description>
-    <Company>Ensono</Company>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/ensono/stacks-dotnet-packages-api-swagger</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/ensono/stacks-dotnet-packages-api-swagger</RepositoryUrl>
-    <RepositoryType>GIT</RepositoryType>
-    <LangVersion>12.0</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.7.0" />
-    <PackageReference Include="System.Text.Encodings.Web" Version="8.0.0" />
-  </ItemGroup>
-
-</Project>
