+++ src/simple-api/src/api/xxENSONOxx.xxSTACKSxx.API/Filters/VersionPathFilter.cs
@@ -1,0 +1,28 @@
+using System.Linq;
+using Microsoft.OpenApi.Models;
+using Swashbuckle.AspNetCore.SwaggerGen;
+
+namespace xxENSONOxx.xxSTACKSxx.API.Filters;
+
+/// <summary>
+/// Filter operation based on path versioning
+/// </summary>
+public class VersionPathFilter(string basePath) : IDocumentFilter
+{
+    public string BasePath { get; } = basePath;
+
+    public void Apply(OpenApiDocument swaggerDoc, DocumentFilterContext context)
+    {
+        if (!BasePath.Contains(swaggerDoc.Info.Version))
+            return;
+
+        var pathsToModify = swaggerDoc.Paths.ToList();
+        foreach (var path in pathsToModify)
+        {
+            if (!path.Key.StartsWith(BasePath))
+            {
+                swaggerDoc.Paths.Remove(path.Key);
+            }
+        }
+    }
+}
