+++ src/simple-api/src/api/xxENSONOxx.xxSTACKSxx.API/xxENSONOxx.xxSTACKSxx.API.csproj
@@ -7,7 +7,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.API.Swagger\xxENSONOxx.xxSTACKSxx.Shared.API.Swagger.csproj" />
         <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.API\xxENSONOxx.xxSTACKSxx.Shared.API.csproj" />
         <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.API.Models\xxENSONOxx.xxSTACKSxx.API.Models.csproj" />
     </ItemGroup>
