+++ xxENSONOxx.xxSTACKSxx.sln
@@ -41,8 +41,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.API.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.API.Tests\xxENSONOxx.xxSTACKSxx.Shared.API.Tests.csproj", "{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.API.Swagger", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.API.Swagger\xxENSONOxx.xxSTACKSxx.Shared.API.Swagger.csproj", "{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Testing", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Testing\xxENSONOxx.xxSTACKSxx.Shared.Testing.csproj", "{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests\xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests.csproj", "{6D545555-2426-4478-82A9-D756A00F1E39}"
@@ -351,18 +349,6 @@ Global
 		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x64.Build.0 = Release|Any CPU
 		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x86.ActiveCfg = Release|Any CPU
 		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x86.Build.0 = Release|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Debug|x64.Build.0 = Debug|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Debug|x86.Build.0 = Debug|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Release|x64.ActiveCfg = Release|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Release|x64.Build.0 = Release|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Release|x86.ActiveCfg = Release|Any CPU
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967}.Release|x86.Build.0 = Release|Any CPU
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -794,7 +780,6 @@ Global
 		{BEE416A3-8EB2-4415-9518-0D0D58EDA7B3} = {A93ABC7D-17EC-4D7E-B8CF-94F11D94910D}
 		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{CCF4C321-DB1D-41D6-AE84-BEA10B1FC967} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{6D545555-2426-4478-82A9-D756A00F1E39} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{DAA32E63-80B5-49B6-890E-7A11FDA6A85B} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
